### 2017-08-01 #343 @whedon
[PRE REVIEW]: prisonbrief: An R package that returns tidy data from the World Prison Brief website
**Submitting author:** @danilofreire (<a href="http://orcid.org/0000-0002-4712-6810">Danilo Freire</a>)
**Repository:** <a href="https://github.com/danilofreire/prisonbrief" target ="_blank">https://github.com/danilofreire/prisonbrief</a>
**Version:** v0.1.0
