### 2017-08-02 #343 @arfon
👋 @openjournals/ropensci - would this be a good candidate for rOpenSci?
