### 2017-08-03 #343 @danilofreire
Thanks @arfon and @maelle for considering the package for rOpenSci. That was very nice of you!

@arfon, I was also reading the list of potential reviewers and I thought that maybe @jsta would be a good fit. He wrote he is comfortable reviewing api wrappers in R, and that's basically what our package does. Also, @benmarwick stated that he has an interest in social sciences, which is the main area of our article. We would be very happy to have them reviewing our code. Of course, we are open to any suggestion you may have. Many thanks!
