### 2017-08-03 #343 @karthik
@danilofreire Thank you for the submission. I read over the paper in the repo and it reads very much like a vignette on how to use the package. Your paper.md must provide a high level description of what the software does and what some of the potential research applications are. You an still have a detailed vignette describing exactly how to use the software.
