+++ paper/paper.bib
@@ -1,0 +1,17 @@
+@article{wickham2014tidy,
+  title={Tidy data},
+  author={Wickham, Hadley},
+  journal={Journal of Statistical Software},
+  volume={59},
+  number={10},
+  pages={1--23},
+  year={2014},
+  publisher={Foundation for Open Access Statistics}
+}
+
+@TechReport{icpr2017prison,
+    title = {World Prison Brief},
+    year = {2017},
+    author ={{Institute for Criminal Policy Research}},
+    editor = {Institute for Criminal Policy Research}, 
+}
