+++ paper/paper.md
@@ -16,113 +16,14 @@ affiliations:
    index: 1
  - name: Avanade
    index: 2
-date: 30 July 2017
+bibliography: paper.bib
+date: 3 August 2017
 ---
 
 # Summary
 
-`prisonbrief` is an R package that downloads, cleans and returns tidy data from the [World Prison Brief website](http://www.prisonstudies.org/). The World Prison Brief is an online database compiled by the [Institute for Criminal Policy Research](http://www.icpr.org.uk/) with information on prison systems around the world. Data currently cover 223 jurisdictions and have been collected from public sources. The `prisonbrief` package provides easy-to-use functions to convert WPB data into a format convenient for statistical analysis. `prisonbrief` has been published to the [Comprehensive Archive Network](https://cran.r-project.org/package=prisonbrief) and has the following DOI: <http://doi.org/10.5281/zenodo.833890>.
+`prisonbrief` is an R package that downloads, cleans and returns tidy data [@wickham2014tidy] from the [World Prison Brief website](http://www.prisonstudies.org/). The World Prison Brief [-@icpr2017prison] is an online database compiled by the [Institute for Criminal Policy Research](http://www.icpr.org.uk/) with information on prison systems around the world. Data currently cover 223 jurisdictions and have been collected from public sources. The `prisonbrief` package provides three easy-to-use functions that convert WPB data into a format convenient for statistical analysis. Researchers interested in comparative prison systems -- such as criminologists, sociologists and economists -- can use the output from functions in the package to create maps, plot time series graphs and calculate descriptive statistics with only a few lines of code.
 
-`prisonbrief` is quite simple to use. The package contains only three functions, all of them starting with `wpb`, a mnemonic for World Prison Brief.
+# References
 
-The first is a convenience function named `wpb_list()`. It prints a list of available countries to the console.
 
-``` r
-library(prisonbrief)
-
-wpb_list()
-#> # A tibble: 226 x 2
-#>                 country_name             country_url
-#>                        <chr>                   <chr>
-#>  1               Afghanistan             afghanistan
-#>  2                   Albania                 albania
-#>  3                   Algeria                 algeria
-#>  4     American Samoa  (USA)      american-samoa-usa
-#>  5                   Andorra                 andorra
-#>  6                    Angola                  angola
-#>  7 Anguilla (United Kingdom) anguilla-united-kingdom
-#>  8       Antigua and Barbuda     antigua-and-barbuda
-#>  9                 Argentina               argentina
-#> 10                   Armenia                 armenia
-#> # ... with 216 more rows
-```
-
-The second function is `wpb_table()`. This function returns a series of variables about the prison systems of the world, of a particular region, or of a specific country. For instance, the code below downloads prison data for Africa:
-
-``` r
-africa <- wpb_table(region = "Africa")
-names(africa)
-#>  [1] "country"                 "prison_population_rate" 
-#>  [3] "prison-population-total" "female-prisoners"       
-#>  [5] "pre-trial-detainees"     "foreign-prisoners"      
-#>  [7] "occupancy-level"         "iso_a2"                 
-#>  [9] "name"                    "geometry"
-```
-
-The region choices are "Africa", "Asia", "Caribbean", "Central America", "Europe", "Middle East", "North America", "Oceania", "South America" and "All".
-
-`wpb_table()` also provides geometric shapes for maps. For instance, you can download and plot the prison population rate in South America with only a few lines of code:
-
-``` r
-south_america <- wpb_table(region = "South America")
-
-library(ggplot2)
-ggplot(south_america, aes(fill = prison_population_rate)) +
-        geom_sf() +
-        scale_fill_distiller(palette = "YlOrRd", trans = "reverse") +
-        theme_minimal()
-```
-
-![](http://i.imgur.com/JxO0wCr.png)
-
[…]
