### 2017-08-03 #343 @danilofreire
@karthik Thank you for your comment. I have just submitted a new version of paper.md. It now includes only a high-level description of the package and some potential applications as requested.
