### 2017-08-11 #361 @whedon
[REVIEW]: prisonbrief: An R package that returns tidy data from the World Prison Brief website
**Submitting author:** @danilofreire (<a href="http://orcid.org/0000-0002-4712-6810">Danilo Freire</a>)
**Repository:** <a href="https://github.com/danilofreire/prisonbrief" target ="_blank">https://github.com/danilofreire/prisonbrief</a>
**Version:** v0.1.0

BM: On [binder/rocker](https://hub.beta.mybinder.org/user/binder-examples-kerfile-rstudio-or9jjiyw/proxy/47529/) I got an error: `ERROR: dependency ‘udunits2’ is not available for package ‘units’`. Maybe add "On Linux you may need to `sudo apt-get install libudunits2-dev` before installing this package"
BM: I see this has been addressed

BM: I don't see any tests
BM:  I see them now
BM: I recommend to add a code of conduct. Could be pasted in to the contributing.md file if you don't want to add another file to the repo.
BM: I see this now.

BM: need to cite source of data in paper.md
BM: I see the citations now
