### 2017-08-12 #361 @benmarwick
I ran

```
> library(prisonbrief)
> wpb_list()
```
and got

```
Error in xpath_class(<S4 object of class "Class">) :
  could not find function "xpath_class"
```
I installed the [selectr](https://github.com/sjp/selectr) pkg and that solved it.

Do we need selectr in the `DESCRIPTION Imports:` ?
