### 2017-08-12 #361 @danilofreire
Hi @benmarwick,

Thanks for pointing that out. I thought that `selectr` would be installed automatically. I'll add the package to `DESCRIPTION Imports:`.
