### 2017-08-12 c7c042d
add selectr, remove downloads/mouth
