### 2017-08-12 8ae4ed7
add selectr, remove downloads/month
