+++ DESCRIPTION
@@ -14,18 +14,19 @@ LazyData: yes
 Depends:
     R (>= 3.0)
 Imports:
-    httr (>= 1.2.1),
-    tibble (>= 1.3.3),
-    dplyr (>= 0.5.0),
     data.table(>= 1.10.4),
+    dplyr (>= 0.5.0),
+    httr (>= 1.2.1),
     magrittr (>= 1.5),
-    rvest (>= 0.3.2),
-    xml2 (>= 1.1.1),
-    tidyr (>= 0.6.3),
-    rlang (>= 0.1.1),
     passport (>= 0.1.1),
+    rlang (>= 0.1.1),
     rnaturalearth (>= 0.1.0),
     rnaturalearthdata (>= 0.1.0),
-    stringr(>= 1.2.0)
+    rvest (>= 0.3.2),
+    selectr (>= 0.3-1),
+    stringr (>= 1.2.0),
+    tibble (>= 1.3.3),
+    tidyr (>= 0.6.3),
+    xml2 (>= 1.1.1)
 RoxygenNote: 6.0.1
 Suggests: testthat
