### 2017-08-13 #361 @benmarwick
Thanks, in the code in  in `R/`,  we see a lot of `::` in functions where `@IimportFrom` is also used in the roxygen above the function. I'm not sure that you need both of these methods. In http://r-pkgs.had.co.nz/namespace.html we see

> If you are using functions repeatedly, you can avoid :: by importing the function with @importFrom pgk fun. This also has a small performance benefit, because :: adds approximately 5 µs to function evaluation time.

So I'd recommend removing the  `::`, what do you think?
