### 2017-08-13 #361 @benmarwick
In the README.Rmd I think we may we need `library(dplyr)` before line 103 to make `%>%` available to the user's environment.
