### 2017-08-13 23aebcb
remove :: from functions, edit readme
