+++ R/wpb_list.R
@@ -9,20 +9,20 @@
 #' @export
 wpb_list <- function(){
 
-  lista <- xml2::read_html("http://www.prisonstudies.org/map/africa") %>%
-    rvest::html_nodes("ul.menu") %>%
-    rvest::html_nodes("div.item-list") %>%
-    rvest::html_nodes("span.field-content") %>%
-    rvest::html_text()
+  lista <- read_html("http://www.prisonstudies.org/map/africa") %>% 
+          html_nodes("ul.menu") %>%
+          html_nodes("div.item-list") %>%
+          html_nodes("span.field-content") %>%
+          html_text()
 
   lista <- sort(unique(lista))
 
   country_url = sort(c("algeria", "angola", "benin", "botswana", "burkina-faso", "burundi", "cameroon", "cape-verde-cabo-verde", "central-african-republic", "chad", "comoros", "congo-brazzaville", "cote-divoire", "democratic-republic-congo-formerly-zaire", "djibouti", "egypt", "equatorial-guinea", "eritrea", "ethiopia", "gabon", "gambia", "ghana", "guinea-bissau", "kenya", "lesotho", "liberia", "libya", "madagascar", "malawi", "mali", "mauritania", "mauritius", "mayotte-france", "morocco", "mozambique", "namibia", "niger", "nigeria", "republic-guinea", "reunion-france", "rwanda", "sao-tome-e-principe", "senegal", "seychelles", "sierra-leone", "somalia", "south-africa", "south-sudan", "sudan", "swaziland", "tanzania", "togo", "tunisia", "uganda", "zambia", "zimbabwe", "afghanistan", "bangladesh", "bhutan", "brunei-darussalam", "cambodia", "china", "democratic-peoples-republic-north-korea", "hong-kong-china", "india", "indonesia", "iran", "japan", "kazakhstan", "kyrgyzstan", "laos", "macau-china", "malaysia", "maldives", "mongolia", "myanmar-formerly-burma", "nepal", "pakistan", "philippines", "republic-south-korea", "singapore", "sri-lanka", "taiwan", "tajikistan", "thailand", "turkmenistan", "uzbekistan", "vietnam", "anguilla-united-kingdom", "antigua-and-barbuda", "aruba-netherlands", "bahamas", "barbados", "cayman-islands-united-kingdom", "cuba", "cura\u00a7ao-netherlands", "dominica", "Dominican Republic", "grenada", "guadeloupe-france", "haiti", "jamaica", "martinique-france", "puerto-rico-usa", "sint-maarten-netherlands", "st-kitts-and-nevis", "st-lucia", "st-vincent-and-grenadines", "trinidad-and-tobago", "virgin-islands-united-kingdom", "virgin-islands-usa", "belize", "costa-rica", "el-salvador", "guatemala", "honduras", "mexico", "nicaragua", "panama", "albania", "andorra", "armenia", "austria", "azerbaijan", "belarus", "belgium", "bosnia-and-herzegovina-federation", "bosnia-and-herzegovina-republika-srpska", "bulgaria", "croatia", "cyprus-republic", "czech-republic", "denmark", "estonia", "faeroe-islands-denmark", "finland", "france", "georgia", "germany", "gibraltar-united-kingdom", "greece", "guernsey-united-kingdom", "hungary", "iceland", "ireland-republic", "isle-man-united-kingdom", "italy", "jersey-united-kingdom)", "kosovokosova", "latvia", "liechtenstein", "lithuania", "luxembourg", "macedonia-former-yugoslav-republic", "malta", "moldova-republic", "monaco", "montenegro", "netherlands", "norway", "poland", "portugal", "romania", "russian-federation", "san-marino", "serbia", "slovakia", "slovenia", "spain", "swede
[…]
