### 2017-08-13 #361 @danilofreire
Thanks again for your comments. We left the `::` in functions because we thought it could improve readability, but you're right in pointing out that they lead to a decrease in performance. All files have been edited now. Also, I've added `library(dplyr)` to `README.md` (line 93 in the new document). You can see the changes [here](https://github.com/danilofreire/prisonbrief/commit/23aebcbf05de70bc035017425b96aa9e6b5669d5).
