### 2017-10-24 #361 @danilofreire
Hi @benmarwick and @karthik! I would like to kindly ask you whether there is anything we should do to continue the reviewing process. My co-author and I are happy to address any further comments the editors may have. Many thanks!
