### 2017-10-26 #361 @karthik
Hi @danilofreire
Sorry for the delay.  @benmarwick will need to check off the items on the list above and give me a green light before I can do a final check and proceed with accepting.

Ben: Can you check off the appropriate boxes? 🙏
