### 2017-11-01 #361 @benmarwick
Sorry to be slow, I've had a go at the checklist, a few easy fixes needed. Let me know when those are done and I'll finish it off.
