+++ paper/paper.bib
@@ -13,5 +13,7 @@
     title = {World Prison Brief},
     year = {2017},
     author ={{Institute for Criminal Policy Research}},
-    editor = {Institute for Criminal Policy Research}, 
+    editor = {Institute for Criminal Policy Research},
+    url = {http://www.prisonstudies.org},
+    note = {Accessed November 2017}
 }
