### 2017-12-05 9eba895
global vars
