+++ R/zz.R
@@ -1,2 +1,2 @@
 
-globalVariables(c(".", "X1", "X2", "X3", "Country"))
+globalVariables(c(".", "X1", "X2", "X3", "Country", "filter", "html_node"))
