### 2017-12-05 #361 @danilofreire
Hi @karthik and @benmarwick. My co-author and I have already addressed the missings items from the checking list. We added the data source as recommended by the original authors and included a message about the Linux dependency [in this commit](https://github.com/danilofreire/prisonbrief/commit/93c748d9e48abba6b70d9e4a4827c0f66efd89a7). Also, we added tests as requested, [please see this commit](https://github.com/danilofreire/prisonbrief/commit/7d605107080ad955d43e9ccbad3c216331f58283). Finally, we included a `Contributing.md` file with instructions on how others can contribute to our project, ask for help, and provide feedback. [Here](https://github.com/danilofreire/prisonbrief/blob/master/CONTRIBUTING.md). We hope these files solve the issues above, and we're happy to address any further questions. Many thanks!
