### 2018-01-23 #361 @danilofreire
Many thanks @benmarwick and @karthik! I greatly appreciate your time and feedback. Please let me or my co-author know if there's anything we should do at the moment. Again, thanks!
