### 2018-01-23 #361 @karthik
I have a few comments myself.

- How come there is `appveyor` integration but none for `travis`?
- [Here]([prisonbrief/wpb_table.R at master · danilofreire/prisonbrief · GitHub](https://github.com/danilofreire/prisonbrief/blob/master/R/wpb_table.R#L65)) you use `html_node` but only import `html_nodes` which throws an error for me. Can you keep it consistent?
- After loading `rvest` as well, I get this error

```
> africa <- wpb_table(region = "Africa")
Error in data.matrix(data) :
  (list) object cannot be coerced to type 'double'
In addition: Warning messages:
1: In data.matrix(data) : NAs introduced by coercion
2: In data.matrix(data) : NAs introduced by coercion
3: In data.matrix(data) : NAs introduced by coercion
```

 <summary>My session info</summary>
<pre>
Session info ------------------------------------------------------------------
 setting  value
 version  R version 3.4.3 (2017-11-30)
 system   x86_64, darwin15.6.0
 ui       RStudio (1.1.383)
 language (EN)
 collate  en_US.UTF-8
 tz       America/Los_Angeles
 date     2018-01-22

Packages ----------------------------------------------------------------------
 package           * version    date
 assertthat          0.2.0      2017-04-11
 base              * 3.4.3      2017-12-07
 bindr               0.1        2016-11-13
 bindrcpp          * 0.2        2017-06-17
 class               7.3-14     2015-08-30
 classInt            0.1-24     2017-04-16
 cli                 1.0.0      2017-11-05
 colorspace          1.3-2      2016-12-14
 compiler            3.4.3      2017-12-07
 coyote              0.3        2017-09-13
 crayon              1.3.4      2017-09-16
 curl                3.1        2017-12-12
 data.table          1.10.4-3   2017-10-27
 datasets          * 3.4.3      2017-12-07
 DBI                 0.7        2017-06-18
 devtools          * 1.13.4     2017-11-09
 digest              0.6.14     2018-01-14
 dplyr               0.7.4      2017-09-28
 e1071               1.6-8      2017-02-02
 ggplot2           * 2.2.1      2016-12-30
 git2r               0.21.0     2018-01-04
 glue                1.2.0      2017-10-29
 graphics          * 3.4.3      2017-12-07
 grDevices         * 3.4.3      2017-12-07
 grid                3.4.3      2017-12-07
 gtable              0.2.0      2016-02-26
 httr                1.3.1      2017-08-20
 lattice             0.20-35    2017-03-25
 lazyeval            0.2.1      2017-10-29
 magrittr            1.5        2014-11-22
 memoise             1.1.0      2017-04-21
 methods           * 3.4.3      2017-12-07
 munsell             0.4.3      2016-02-13
 passport            0.2.0      2017-11-30
 pillar              1.1.0      2018-01-14
 pkgconfig           2.0.1      2017-03-21
 plyr                1.8.4      2016-06-08
 prisonbrief       * 0.1.0      2018-01-23
 purrr               0.2.4      2017-10-18
 R6                  2.2.2      2017-06-17
 Rcpp                0.12.15    2018-01-20
 rgeos               0.3-26     2017-10-31
 rlang               0.1.6      2017-12-21
 rnaturalearth       0.1.0      2017-03-21
 rnaturalearthdata   0.1.0      2017-02-21
 rstudioapi          0.7.0-9000 2017-10-05
 rvest             * 0.3.2      2016-06-17
 scales              0.5.0      2017-08-24
 selectr             0.3-1      2016-12-19
 sf                  0.6-0      2018-01-06
 sp                  1.2-7      2018-01-19
 stats             * 3.4.3      2017-12-07
 stringi             1.1.6      2017-11-17
 stringr             1.2.0      2017-02-18
 tibble              1.4.1      2017-12-25
 tidyr               0.7.2      2017-10-16
 tools               3.4.3      2017-12-07
 udunits2            0.13       2016-11-17
 units               0.5-1      2018-01-08
 utf8                1.1.3      2018-01-03
 utils             * 3.4.3      2017-12-07
 withr               2.1.1      2017-12-19
 XML                 3.98-1.9   2017-06-19
 xml2              * 1.1.1      2017-01-24
 yaml                2.1.16     2017-12-12
 source
 CRAN (R 3.4.0)
 local
 CRAN (R 3.4.0)
 CRAN (R 3.4.0)
 CRAN (R 3.4.3)
 CRAN (R 3.4.0)
 CRAN (R 3.4.2)
 CRAN (R 3.4.0)
 local
 Github (karthik/coyote@acad466)
 cran (@1.3.4)
 CRAN (R 3.4.3)
 CRAN (R 3.4.2)
 local
 CRAN (R 3.4.0)
 CRAN (R 3.4.2)
 CRAN (R 3.4.3)
 CRAN (R 3.4.2)
 CRAN (R 3.4.0)
 CRAN (R 3.4.0)
 CRAN (R 3.4.2)
 CRAN (R 3.4.1)
 local
 local
 local
 CRAN (R 3.4.0)
 CRAN (R 3.4.1)
 CRAN (R 3.4.3)
 CRAN (R 3.4.1)
 CRAN (R 3.4.0)
 CRAN (R 3.4.0)
 local
 CRAN (R 3.4.0)
 cran (@0.2.0)
 CRAN (R 3.4.3)
 CRAN (R 3.4.0)
 CRAN (R 3.4.0)
 Github (danilofreire/prisonbrief@9eba895)
 CRAN (R 3.4.2)
 CRAN (R 3.4.0)
 CRAN (R 3.4.3)
 CRAN (R 3.4.2)
 CRAN (R 3.4.3)
 cran (@0.1.0)
 cran (@0.1.0)
 Github (rstudio/rstudioapi@335f257)
 CRAN (R 3.4.0)
 CRAN (R 3.4.1)
 CRAN (R 3.4.0)
 CRAN (R 3.4.3)
 CRAN (R 3.4.3)
 local
 CRAN (R 3.4.2)
 CRAN (R 3.4.0)
 CRAN (R 3.4.2)
 CRAN (R 3.4.2)
 local
 CRAN (R 3.4.0)
 CRAN (R 3.4.2)
 CRAN (R 3.4.2)
 local
 CRAN (R 3.4.2)
 CRAN (R 3.4.1)
 CRAN (R 3.4.0)
 CRAN (R 3.4.3)
</pre>
