### 2018-01-26 df6e622
put back :: in dplyr::filter/select, fixed error on column names for countries in wpb_table() :es:
