+++ R/zz.R
@@ -1,2 +1,12 @@
+#' Pipe operator
+#'
+#' @name %>%
+#' @rdname pipe
+#' @keywords internal
+#' @export
+#' @importFrom magrittr %>%
+#' @usage lhs \%>\% rhs
+NULL
 
-globalVariables(c(".", "X1", "X2", "X3", "Country", "filter", "html_node"))
+globalVariables(c(".", "X1", "X2", "X3", "Country", "filter", "html_node", "name", 
+                  "iso_a2", "Ministry responsible", "Prison administration"))
