+++ NAMESPACE
@@ -1,5 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+export("%>%")
 export(wbp_region_list)
 export(wpb_list)
 export(wpb_series)
@@ -7,6 +8,7 @@ export(wpb_table)
 import(rnaturalearthdata)
 importFrom(data.table,':=')
 importFrom(dplyr,bind_cols)
+importFrom(dplyr,contains)
 importFrom(dplyr,everything)
 importFrom(dplyr,full_join)
 importFrom(dplyr,funs)
@@ -16,12 +18,12 @@ importFrom(dplyr,select)
 importFrom(dplyr,slice)
 importFrom(httr,GET)
 importFrom(magrittr,"%>%")
-importFrom(magrittr,'%>%')
 importFrom(magrittr,set_colnames)
 importFrom(passport,as_country_name)
 importFrom(passport,parse_country)
 importFrom(rlang,UQ)
 importFrom(rnaturalearth,ne_countries)
+importFrom(rvest,html_node)
 importFrom(rvest,html_nodes)
 importFrom(rvest,html_table)
 importFrom(rvest,html_text)
