### 2018-01-29 098fb32
removed data.table, modified wpb_table and included rlang ':=' :es:
