+++ NAMESPACE
@@ -6,7 +6,6 @@ export(wpb_list)
 export(wpb_series)
 export(wpb_table)
 import(rnaturalearthdata)
-importFrom(data.table,':=')
 importFrom(dplyr,bind_cols)
 importFrom(dplyr,contains)
 importFrom(dplyr,everything)
@@ -21,6 +20,7 @@ importFrom(magrittr,"%>%")
 importFrom(magrittr,set_colnames)
 importFrom(passport,as_country_name)
 importFrom(passport,parse_country)
+importFrom(rlang,':=')
 importFrom(rlang,UQ)
 importFrom(rnaturalearth,ne_countries)
 importFrom(rvest,html_node)
