### 2018-01-29 5a21152
:albania: fixed wrongly typed function name and error in returned data
