+++ man/wpb_region_list.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/wpb_list.R
-\name{wbp_region_list}
-\alias{wbp_region_list}
+\name{wpb_region_list}
+\alias{wpb_region_list}
 \title{Lists all countries in a region}
 \usage{
-wbp_region_list(region = c("Africa", "Asia", "Caribbean", "Central America",
+wpb_region_list(region = c("Africa", "Asia", "Caribbean", "Central America",
   "Europe", "Middle East", "North America", "Oceania", "South America"))
 }
 \arguments{
