+++ NAMESPACE
@@ -1,8 +1,8 @@
 # Generated by roxygen2: do not edit by hand
 
 export("%>%")
-export(wbp_region_list)
 export(wpb_list)
+export(wpb_region_list)
 export(wpb_series)
 export(wpb_table)
 import(rnaturalearthdata)
