+++ NAMESPACE
@@ -13,6 +13,7 @@ importFrom(dplyr,full_join)
 importFrom(dplyr,funs)
 importFrom(dplyr,mutate)
 importFrom(dplyr,mutate_all)
+importFrom(dplyr,rename)
 importFrom(dplyr,select)
 importFrom(dplyr,slice)
 importFrom(httr,GET)
@@ -20,8 +21,6 @@ importFrom(magrittr,"%>%")
 importFrom(magrittr,set_colnames)
 importFrom(passport,as_country_name)
 importFrom(passport,parse_country)
-importFrom(rlang,':=')
-importFrom(rlang,UQ)
 importFrom(rnaturalearth,ne_countries)
 importFrom(rvest,html_node)
 importFrom(rvest,html_nodes)
