+++ DESCRIPTION
@@ -14,12 +14,10 @@ LazyData: yes
 Depends:
     R (>= 3.0)
 Imports:
-    data.table(>= 1.10.4),
     dplyr (>= 0.5.0),
     httr (>= 1.2.1),
     magrittr (>= 1.5),
     passport (>= 0.1.1),
-    rlang (>= 0.1.1),
     rnaturalearth (>= 0.1.0),
     rnaturalearthdata (>= 0.1.0),
     rvest (>= 0.3.2),
