+++ R/zz.R
@@ -9,4 +9,5 @@
 NULL
 
 globalVariables(c(".", "X1", "X2", "X3", "Country", "filter", "html_node", "name", 
-                  "iso_a2", "Ministry responsible", "Prison administration"))
+                  "iso_a2", "Ministry responsible", "Prison administration",
+                  "Ranking", "Title"))
