+++ DESCRIPTION
@@ -13,6 +13,7 @@ Encoding: UTF-8
 LazyData: yes
 Depends:
     R (>= 3.0)
+Remotes: r-spatial/sf
 Imports:
     dplyr (>= 0.5.0),
     httr (>= 1.2.1),
