+++ DESCRIPTION
@@ -15,6 +15,7 @@ Depends:
     R (>= 3.0)
 Remotes: r-spatial/sf
 Imports:
+    sf (>= 0.6.0),
     dplyr (>= 0.5.0),
     httr (>= 1.2.1),
     magrittr (>= 1.5),
