+++ DESCRIPTION
@@ -16,8 +16,6 @@ Depends:
 Remotes: r-spatial/sf
 Imports:
     sf (>= 0.6.0),
-    units (>= 0.4.6),
-    udunits2 (>= 0.13),
     dplyr (>= 0.5.0),
     httr (>= 1.2.1),
     magrittr (>= 1.5),
