+++ .travis.yml
@@ -1,5 +1,9 @@
 # R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
 
+before_install:
+- sudo apt-get -qq update
+- sudo apt-get install -y libudunits2-dev
+
 language: R
 sudo: false
 cache: packages
