+++ .travis.yml
@@ -2,7 +2,7 @@
 
 before_install:
 - sudo apt-get -qq update
-- sudo apt-get install -y libudunits2-dev
+- sudo apt-get install -y libudunits2-dev libgdal-dev libgeos-dev libproj-dev 
 
 language: R
 sudo: false
