+++ .travis.yml
@@ -1,52 +1,10 @@
 # R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
 
+before_install:
+- sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable
+- sudo apt-get -qq update
+- sudo apt-get install libudunits2-dev libgdal-dev libgeos-dev libproj-dev 
+
 language: r
 cache: packages
-
-apt_packages:
-  - libudunits2-dev
-
-addons:
-  postgresql: "9.3"
-
-before_script:
-  - psql -U postgres -c "create extension postgis;"
-
-before_install:
-  - createdb postgis
-  - psql -d postgis -c "CREATE EXTENSION postgis;"
-  - createdb empty
-  - psql -d empty -c "CREATE EXTENSION postgis;"
-  - export CURWD=`pwd`
-  - sudo apt-get install libproj-dev libgeos++-dev
-  - cd
-  - wget http://download.osgeo.org/gdal/2.1.0/gdal-2.1.0.tar.gz
-  - tar zxf gdal-2.1.0.tar.gz
-  - cd gdal-2.1.0
-  - ./configure
-  - make
-  - sudo make install
-  - cd
-  - wget http://download.osgeo.org/proj/proj-4.9.3.tar.gz
-  - tar zxvf proj-4.9.3.tar.gz
-  - cd proj-4.9.3
-  - ./configure
-  - make
-  - sudo make install
-  - sudo ldconfig
-  - cd $CURWD
-
 sudo: required
-
-warnings_are_errors: true
-
-r_packages:
-  - rgdal
-  - roxygen2
-  - rmarkdown
-
-r_github_packages: 
-  - jimhester/covr
-
-after_success:
-  - R -e 'covr::codecov()'
