+++ .travis.yml
@@ -1,7 +1,7 @@
 # R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
 
 before_install:
-- sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable
+- sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable -y
 - sudo apt-get -qq update
 - sudo apt-get install libudunits2-dev libgdal-dev libgeos-dev libproj-dev 
 
