### 2018-02-03 #361 @danilofreire
Hi @karthik, thanks very much for your comments. We've fixed the errors you pointed out.

* `travis` is now integrated: [https://travis-ci.org/danilofreire/prisonbrief](https://travis-ci.org/danilofreire/prisonbrief)
* the script is now consistent and should not throw an error anymore.
 See: [https://github.com/danilofreire/prisonbrief/blob/master/R/wpb_table.R](https://github.com/danilofreire/prisonbrief/blob/master/R/wpb_table.R)
* we fixed the problem with region names with this commit: [https://github.com/danilofreire/prisonbrief/commit/5a21152c5154f83af0928a69e1e54f82de75f64f](https://github.com/danilofreire/prisonbrief/commit/5a21152c5154f83af0928a69e1e54f82de75f64f).

Please let us know if you encounter any problem with the code. Many thanks!
