+++ NEWS.md
@@ -1,0 +1,9 @@
+# prisonbrief 0.1.1
+
+* Added a `NEWS.md` file to track changes to the package.
+* Fixed some small errors about region names.
+* Small changes to the code to increase performance.
+* More details about the fixes [here](https://github.com/openjournals/joss-reviews/issues/361).
+
+
+
